requirements.txt
apache-airflow==2.9.1
dbt==1.0.0.37.18
dbt-snowflake==1.8.2
requests==2.32.3
snowflake-connector-python[pandas]==3.10.1
snowflake-sqlalchemy==1.5.3
pytest==8.2.2
meteomatics==2.11.2
pandas==2.2.2
